Styler-OP-backend/pyproject.toml
[project]
name = "styler-op-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Arpan Mondal",email = "arpanmondal572@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi (>=0.124.2,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pymongo[srv] (>=4.15.5,<5.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib (==1.7.4)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "motor (>=3.7.1,<4.0.0)",
    "certifi (>=2025.11.12,<2026.0.0)",
    "bcrypt (==4.1.3)",
    "argon2-cffi (>=25.1.0,<26.0.0)"
]
[tool.poetry]
package-mode = false 

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
